In order to make sent_subfolders work I had to change a lot of code. I may have messed up something in the process so PLEASE TEST THIS CARFULLY.
In my system (that uses uw) it almost work. The only thing left to
be done is to organize left main in such a way that sent subfolders
are placed to the top.

I've replaced some spare code with a few new functions.
sqimap_mailbox_rename and isSpecialMailbox are the most interesting
in other proceses.


git-svn-id: svn://svn.code.sf.net/p/squirrelmail/code@2119 7612ce4b-ef26-0410-bec9-ea0150e637f0