Some tweaks to SMTP auth to allow a site-wide user/pass override (for those of us with a home installation that want to go out using the ISP's SMTP server because people arbitrarily block our dynamic IP addresses).
new function in auth.php assigns the correct user/pass values
based on the auth mechanism.

For some reason the initStream method in Deliver_SMTP was not using
the passed in user name, it was using the global $username, and had
the key and onetimepad declared, though it always used the passed in
parameter pass instead - cleaned that up so it used the user/pass values
passed in instead.

read_body and compose were updated to use the new function in auth.php
to get the correct username and password for SMTP auth, and pass these
values to initStream.

Works for me.


git-svn-id: svn://svn.code.sf.net/p/squirrelmail/code@6369 7612ce4b-ef26-0410-bec9-ea0150e637f0